Updating second implementation
Now the second implementation works for more than 1 author. If the first author does not have 10 articles, it will get the remaining articles from the next author, and so on. Improved the warning messages to show which author it couldn't find in google scholar, and if an author had less than 10 entries in google scholar, now shows how many entries were added.